@@ -1,7 +1,8 @@
 {
   "private": true,
   "workspaces": [
-    "packages/*"
+    "packages/*",
+    "docusaurus/website"
   ],
   "scripts": {
     "build": "cd packages/react-scripts && node bin/react-scripts.js build",
@@ -9,40 +10,42 @@
     "create-react-app": "node tasks/cra.js",
     "e2e": "tasks/e2e-simple.sh",
     "e2e:docker": "tasks/local-test.sh",
-    "postinstall": "cd packages/react-error-overlay/ && yarn build:prod",
+    "postinstall": "npm run build:prod -w react-error-overlay",
     "publish": "tasks/publish.sh",
     "start": "cd packages/react-scripts && node bin/react-scripts.js start",
     "screencast": "node ./tasks/screencast.js",
     "screencast:error": "svg-term --cast jyu19xGl88FQ3poMY8Hbmfw8y --out screencast-error.svg --window --at 12000 --no-cursor",
     "alex": "alex .",
+    "test:integration": "jest test/integration",
     "test": "cd packages/react-scripts && node bin/react-scripts.js test",
-    "format": "prettier --write 'packages/*/*.js' 'packages/*/!(node_modules)/**/*.js'",
-    "compile:lockfile": "node tasks/compile-lockfile.js"
+    "eslint": "eslint .",
+    "prettier": "prettier .",
+    "format": "npm run prettier -- --write"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.9.0",
-    "@testing-library/react": "^10.2.1",
-    "@testing-library/user-event": "^11.3.2",
-    "alex": "^8.0.0",
-    "eslint": "^6.1.0",
-    "execa": "1.0.0",
-    "fs-extra": "^9.0.0",
+    "@testing-library/jest-dom": "^5.15.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "alex": "^8.2.0",
+    "eslint": "^8.3.0",
+    "execa": "^5.1.1",
+    "fs-extra": "^10.0.0",
     "get-port": "^5.1.1",
-    "globby": "^11.0.0",
-    "husky": "^4.2.5",
-    "jest": "26.0.1",
-    "lerna": "3.20.2",
-    "lerna-changelog": "~0.8.2",
-    "lint-staged": "^10.2.2",
-    "meow": "^6.1.1",
-    "multimatch": "^4.0.0",
-    "prettier": "2.0.5",
-    "puppeteer": "^3.0.2",
-    "strip-ansi": "^6.0.0",
+    "globby": "^11.0.4",
+    "husky": "^4.3.8",
+    "jest": "^27.4.3",
+    "lerna": "^4.0.0",
+    "lerna-changelog": "^2.2.0",
+    "lint-staged": "^12.1.2",
+    "meow": "^9.0.0",
+    "multimatch": "^5.0.0",
+    "prettier": "^2.5.0",
+    "puppeteer": "^12.0.1",
+    "strip-ansi": "^6.0.1",
     "svg-term-cli": "^2.1.1",
-    "tempy": "^0.2.1",
-    "wait-for-localhost": "^3.1.0",
-    "web-vitals": "^0.2.2"
+    "tempy": "^1.0.1",
+    "wait-for-localhost": "^3.3.0",
+    "web-vitals": "^2.1.2"
   },
   "husky": {
     "hooks": {
@@ -50,7 +53,7 @@
     }
   },
   "lint-staged": {
-    "*.{js,md,css,html}": [
+    "*.{js,json,yml,yaml,css,scss,ts,tsx,md}": [
       "prettier --write"
     ]
   }
