@@ -1,6 +1,6 @@
 {
   "name": "create-react-app",
-  "version": "3.4.1",
+  "version": "5.0.1",
   "keywords": [
     "react"
   ],
@@ -12,30 +12,36 @@
   },
   "license": "MIT",
   "engines": {
-    "node": ">=10"
+    "node": ">=14"
   },
   "bugs": {
     "url": "https://github.com/facebook/create-react-app/issues"
   },
   "files": [
     "index.js",
-    "createReactApp.js",
-    "yarn.lock.cached"
+    "createReactApp.js"
   ],
   "bin": {
     "create-react-app": "./index.js"
   },
+  "scripts": {
+    "test": "cross-env FORCE_COLOR=true jest"
+  },
   "dependencies": {
-    "chalk": "4.1.0",
-    "commander": "4.1.0",
-    "cross-spawn": "7.0.3",
-    "envinfo": "7.5.1",
-    "fs-extra": "9.0.1",
-    "hyperquest": "2.1.3",
-    "inquirer": "7.1.0",
-    "semver": "7.3.2",
-    "tar-pack": "3.4.1",
-    "tmp": "0.2.1",
-    "validate-npm-package-name": "3.0.0"
+    "chalk": "^4.1.2",
+    "commander": "^4.1.1",
+    "cross-spawn": "^7.0.3",
+    "envinfo": "^7.8.1",
+    "fs-extra": "^10.0.0",
+    "hyperquest": "^2.1.3",
+    "prompts": "^2.4.2",
+    "semver": "^7.3.5",
+    "tar-pack": "^3.4.1",
+    "tmp": "^0.2.1",
+    "validate-npm-package-name": "^3.0.0"
+  },
+  "devDependencies": {
+    "cross-env": "^7.0.3",
+    "jest": "^27.4.3"
   }
 }
