@@ -1,6 +1,6 @@
 {
   "name": "eslint-config-react-app",
-  "version": "5.2.1",
+  "version": "7.0.1",
   "description": "ESLint configuration used by Create React App",
   "repository": {
     "type": "git",
@@ -12,20 +12,30 @@
     "url": "https://github.com/facebook/create-react-app/issues"
   },
   "files": [
-    "index.js"
+    "base.js",
+    "index.js",
+    "jest.js"
   ],
   "peerDependencies": {
-    "@typescript-eslint/eslint-plugin": "2.x",
-    "@typescript-eslint/parser": "2.x",
-    "babel-eslint": "10.x",
-    "eslint": "6.x",
-    "eslint-plugin-flowtype": "3.x || 4.x",
-    "eslint-plugin-import": "2.x",
-    "eslint-plugin-jsx-a11y": "6.x",
-    "eslint-plugin-react": "7.x",
-    "eslint-plugin-react-hooks": "1.x || 2.x || 3.x || 4.x"
+    "eslint": "^8.0.0"
   },
   "dependencies": {
-    "confusing-browser-globals": "^1.0.9"
+    "@babel/core": "^7.16.0",
+    "@babel/eslint-parser": "^7.16.3",
+    "@rushstack/eslint-patch": "^1.1.0",
+    "@typescript-eslint/eslint-plugin": "^5.5.0",
+    "@typescript-eslint/parser": "^5.5.0",
+    "babel-preset-react-app": "^10.0.1",
+    "confusing-browser-globals": "^1.0.11",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jest": "^25.3.0",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-testing-library": "^5.0.1"
+  },
+  "engines": {
+    "node": ">=14.0.0"
   }
 }
