@@ -1,6 +1,6 @@
 {
   "name": "react-app-polyfill",
-  "version": "1.0.6",
+  "version": "3.0.0",
   "description": "Polyfills for various browsers including commonly used language features",
   "repository": {
     "type": "git",
@@ -12,7 +12,7 @@
     "url": "https://github.com/facebook/create-react-app/issues"
   },
   "engines": {
-    "node": ">=10"
+    "node": ">=14"
   },
   "files": [
     "ie9.js",
@@ -21,11 +21,11 @@
     "stable.js"
   ],
   "dependencies": {
-    "core-js": "^3.5.0",
+    "core-js": "^3.19.2",
     "object-assign": "^4.1.1",
-    "promise": "^8.0.3",
+    "promise": "^8.1.0",
     "raf": "^3.4.1",
-    "regenerator-runtime": "^0.13.3",
-    "whatwg-fetch": "^3.0.0"
+    "regenerator-runtime": "^0.13.9",
+    "whatwg-fetch": "^3.6.2"
   }
 }
