@@ -1,6 +1,6 @@
 {
   "name": "react-dev-utils",
-  "version": "10.2.1",
+  "version": "12.0.1",
   "description": "webpack utilities used by Create React App",
   "repository": {
     "type": "git",
@@ -12,7 +12,7 @@
     "url": "https://github.com/facebook/create-react-app/issues"
   },
   "engines": {
-    "node": ">=10"
+    "node": ">=14"
   },
   "files": [
     "browsersHelper.js",
@@ -25,6 +25,7 @@
     "evalSourceMapMiddleware.js",
     "FileSizeReporter.js",
     "ForkTsCheckerWebpackPlugin.js",
+    "ForkTsCheckerWarningWebpackPlugin.js",
     "formatWebpackMessages.js",
     "getCacheIdentifier.js",
     "getCSSModuleLocalIdent.js",
@@ -34,7 +35,6 @@
     "ignoredFiles.js",
     "immer.js",
     "InlineChunkHtmlPlugin.js",
-    "inquirer.js",
     "InterpolateHtmlPlugin.js",
     "launchEditor.js",
     "launchEditorEndpoint.js",
@@ -46,40 +46,40 @@
     "printBuildError.js",
     "printHostingInstructions.js",
     "redirectServedPathMiddleware.js",
+    "refreshOverlayInterop.js",
     "typescriptFormatter.js",
-    "WatchMissingNodeModulesPlugin.js",
     "WebpackDevServerUtils.js",
     "webpackHotDevClient.js"
   ],
   "dependencies": {
-    "@babel/code-frame": "7.10.1",
-    "address": "1.1.2",
-    "browserslist": "4.12.0",
-    "chalk": "2.4.2",
-    "cross-spawn": "7.0.3",
-    "detect-port-alt": "1.1.6",
-    "escape-string-regexp": "2.0.0",
-    "filesize": "6.1.0",
-    "find-up": "4.1.0",
-    "fork-ts-checker-webpack-plugin": "4.1.6",
-    "global-modules": "2.0.0",
-    "globby": "11.0.1",
-    "gzip-size": "5.1.1",
-    "immer": "1.10.0",
-    "inquirer": "7.1.0",
-    "is-root": "2.1.0",
-    "loader-utils": "2.0.0",
-    "open": "^7.0.2",
-    "pkg-up": "3.1.0",
-    "react-error-overlay": "^6.0.7",
-    "recursive-readdir": "2.2.2",
-    "shell-quote": "1.7.2",
-    "strip-ansi": "6.0.0",
-    "text-table": "0.2.0"
+    "@babel/code-frame": "^7.16.0",
+    "address": "^1.1.2",
+    "browserslist": "^4.18.1",
+    "chalk": "^4.1.2",
+    "cross-spawn": "^7.0.3",
+    "detect-port-alt": "^1.1.6",
+    "escape-string-regexp": "^4.0.0",
+    "filesize": "^8.0.6",
+    "find-up": "^5.0.0",
+    "fork-ts-checker-webpack-plugin": "^6.5.0",
+    "global-modules": "^2.0.0",
+    "globby": "^11.0.4",
+    "gzip-size": "^6.0.0",
+    "immer": "^9.0.7",
+    "is-root": "^2.1.0",
+    "loader-utils": "^3.2.0",
+    "open": "^8.4.0",
+    "pkg-up": "^3.1.0",
+    "prompts": "^2.4.2",
+    "react-error-overlay": "^6.0.11",
+    "recursive-readdir": "^2.2.2",
+    "shell-quote": "^1.7.3",
+    "strip-ansi": "^6.0.1",
+    "text-table": "^0.2.0"
   },
   "devDependencies": {
-    "cross-env": "^7.0.2",
-    "jest": "26.0.1"
+    "cross-env": "^7.0.3",
+    "jest": "^27.4.3"
   },
   "scripts": {
     "test": "cross-env FORCE_COLOR=true jest"
