@@ -1,6 +1,6 @@
 {
   "name": "react-error-overlay",
-  "version": "6.0.7",
+  "version": "6.0.11",
   "description": "An overlay for displaying stack frames.",
   "main": "lib/index.js",
   "sideEffects": false,
@@ -8,7 +8,8 @@
     "start": "cross-env NODE_ENV=development node build.js --watch",
     "test": "cross-env NODE_ENV=test jest",
     "build": "cross-env NODE_ENV=development node build.js",
-    "build:prod": "cross-env NODE_ENV=production node build.js"
+    "build:prod": "cross-env NODE_ENV=production node build.js",
+    "flow": "flow"
   },
   "repository": {
     "type": "git",
@@ -34,37 +35,29 @@
     "lib/index.js"
   ],
   "devDependencies": {
-    "@babel/code-frame": "7.10.1",
-    "@babel/core": "7.10.2",
-    "anser": "1.4.9",
-    "babel-eslint": "10.1.0",
-    "babel-jest": "^26.0.1",
-    "babel-loader": "8.1.0",
-    "babel-preset-react-app": "^9.1.2",
-    "chalk": "2.4.2",
-    "chokidar": "^3.3.0",
-    "cross-env": "7.0.2",
-    "eslint": "^6.1.0",
-    "eslint-config-react-app": "^5.2.1",
-    "eslint-plugin-flowtype": "4.7.0",
-    "eslint-plugin-import": "2.20.2",
-    "eslint-plugin-jsx-a11y": "6.2.3",
-    "eslint-plugin-react": "7.20.0",
+    "@babel/code-frame": "^7.16.0",
+    "@babel/core": "^7.16.0",
+    "anser": "^2.1.0",
+    "babel-jest": "^27.4.2",
+    "babel-loader": "^8.2.3",
+    "babel-preset-react-app": "^10.0.1",
+    "chalk": "^4.1.2",
+    "chokidar": "^3.5.2",
+    "cross-env": "^7.0.3",
     "flow-bin": "^0.116.0",
-    "html-entities": "1.3.1",
-    "jest": "26.0.1",
-    "jest-fetch-mock": "2.1.2",
-    "object-assign": "4.1.1",
-    "promise": "8.1.0",
-    "raw-loader": "^4.0.1",
-    "react": "^16.12.0",
-    "react-app-polyfill": "^1.0.6",
-    "react-dom": "^16.12.0",
-    "rimraf": "^3.0.0",
-    "settle-promise": "1.0.0",
-    "source-map": "0.5.6",
-    "terser-webpack-plugin": "3.0.3",
-    "webpack": "^4.35.0"
+    "html-entities": "^2.3.2",
+    "jest": "^27.4.3",
+    "jest-fetch-mock": "^3.0.3",
+    "object-assign": "^4.1.1",
+    "promise": "^8.1.0",
+    "raw-loader": "^4.0.2",
+    "react": "^17.0.2",
+    "react-app-polyfill": "^3.0.0",
+    "react-dom": "^17.0.2",
+    "rimraf": "^3.0.2",
+    "settle-promise": "^1.0.0",
+    "source-map": "^0.5.7",
+    "webpack": "^5.64.4"
   },
   "jest": {
     "setupFiles": [
