@@ -5,6 +5,6 @@
     "@types/jest": "*",
     "react": "*",
     "react-dom": "*",
-    "typescript": "3.7.2"
+    "typescript": "^3.9.5"
   }
 }
