@@ -9,11 +9,6 @@
 //   /!\ DO NOT MODIFY THIS FILE /!\
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 //
-// create-react-app is installed globally on people's computers. This means
-// that it is extremely difficult to have them upgrade the version and
-// because there's only one global version installed, it is very prone to
-// breaking changes.
-//
 // The only job of create-react-app is to init the repository and then
 // forward all the commands to the local version of create-react-app.
 //
@@ -26,7 +21,7 @@
 // tell people to update their global version of create-react-app.
 //
 // Also be careful with new language features.
-// This file must work on Node 6+.
+// This file must work on Node 10+.
 //
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 //   /!\ DO NOT MODIFY THIS FILE /!\
@@ -34,6 +29,7 @@
 
 'use strict';
 
+const https = require('https');
 const chalk = require('chalk');
 const commander = require('commander');
 const dns = require('dns');
@@ -41,7 +37,7 @@ const envinfo = require('envinfo');
 const execSync = require('child_process').execSync;
 const fs = require('fs-extra');
 const hyperquest = require('hyperquest');
-const inquirer = require('inquirer');
+const prompts = require('prompts');
 const os = require('os');
 const path = require('path');
 const semver = require('semver');
@@ -53,148 +49,201 @@ const validateProjectName = require('validate-npm-package-name');
 
 const packageJson = require('./package.json');
 
+function isUsingYarn() {
+  return (process.env.npm_config_user_agent || '').indexOf('yarn') === 0;
+}
+
 let projectName;
 
-const program = new commander.Command(packageJson.name)
-  .version(packageJson.version)
-  .arguments('<project-directory>')
-  .usage(`${chalk.green('<project-directory>')} [options]`)
-  .action(name => {
-    projectName = name;
-  })
-  .option('--verbose', 'print additional logs')
-  .option('--info', 'print environment debug info')
-  .option(
-    '--scripts-version <alternative-package>',
-    'use a non-standard version of react-scripts'
-  )
-  .option(
-    '--template <path-to-template>',
-    'specify a template for the created project'
-  )
-  .option('--use-npm')
-  .option('--use-pnp')
-  .allowUnknownOption()
-  .on('--help', () => {
-    console.log(`    Only ${chalk.green('<project-directory>')} is required.`);
-    console.log();
-    console.log(
-      `    A custom ${chalk.cyan('--scripts-version')} can be one of:`
-    );
-    console.log(`      - a specific npm version: ${chalk.green('0.8.2')}`);
-    console.log(`      - a specific npm tag: ${chalk.green('@next')}`);
-    console.log(
-      `      - a custom fork published on npm: ${chalk.green(
-        'my-react-scripts'
-      )}`
-    );
-    console.log(
-      `      - a local path relative to the current working directory: ${chalk.green(
-        'file:../my-react-scripts'
-      )}`
-    );
-    console.log(
-      `      - a .tgz archive: ${chalk.green(
-        'https://mysite.com/my-react-scripts-0.8.2.tgz'
-      )}`
-    );
-    console.log(
-      `      - a .tar.gz archive: ${chalk.green(
-        'https://mysite.com/my-react-scripts-0.8.2.tar.gz'
-      )}`
-    );
-    console.log(
-      `    It is not needed unless you specifically want to use a fork.`
-    );
-    console.log();
-    console.log(`    A custom ${chalk.cyan('--template')} can be one of:`);
-    console.log(
-      `      - a custom template published on npm: ${chalk.green(
-        'cra-template-typescript'
-      )}`
-    );
-    console.log(
-      `      - a local path relative to the current working directory: ${chalk.green(
-        'file:../my-custom-template'
-      )}`
-    );
-    console.log(
-      `      - a .tgz archive: ${chalk.green(
-        'https://mysite.com/my-custom-template-0.8.2.tgz'
-      )}`
-    );
-    console.log(
-      `      - a .tar.gz archive: ${chalk.green(
-        'https://mysite.com/my-custom-template-0.8.2.tar.gz'
-      )}`
-    );
-    console.log();
-    console.log(
-      `    If you have any problems, do not hesitate to file an issue:`
-    );
-    console.log(
-      `      ${chalk.cyan(
-        'https://github.com/facebook/create-react-app/issues/new'
-      )}`
-    );
-    console.log();
-  })
-  .parse(process.argv);
-
-if (program.info) {
-  console.log(chalk.bold('\nEnvironment Info:'));
-  console.log(
-    `\n  current version of ${packageJson.name}: ${packageJson.version}`
-  );
-  console.log(`  running from ${__dirname}`);
-  return envinfo
-    .run(
-      {
-        System: ['OS', 'CPU'],
-        Binaries: ['Node', 'npm', 'Yarn'],
-        Browsers: ['Chrome', 'Edge', 'Internet Explorer', 'Firefox', 'Safari'],
-        npmPackages: ['react', 'react-dom', 'react-scripts'],
-        npmGlobalPackages: ['create-react-app'],
-      },
-      {
-        duplicates: true,
-        showNotFound: true,
-      }
+function init() {
+  const program = new commander.Command(packageJson.name)
+    .version(packageJson.version)
+    .arguments('<project-directory>')
+    .usage(`${chalk.green('<project-directory>')} [options]`)
+    .action(name => {
+      projectName = name;
+    })
+    .option('--verbose', 'print additional logs')
+    .option('--info', 'print environment debug info')
+    .option(
+      '--scripts-version <alternative-package>',
+      'use a non-standard version of react-scripts'
     )
-    .then(console.log);
+    .option(
+      '--template <path-to-template>',
+      'specify a template for the created project'
+    )
+    .option('--use-pnp')
+    .allowUnknownOption()
+    .on('--help', () => {
+      console.log(
+        `    Only ${chalk.green('<project-directory>')} is required.`
+      );
+      console.log();
+      console.log(
+        `    A custom ${chalk.cyan('--scripts-version')} can be one of:`
+      );
+      console.log(`      - a specific npm version: ${chalk.green('0.8.2')}`);
+      console.log(`      - a specific npm tag: ${chalk.green('@next')}`);
+      console.log(
+        `      - a custom fork published on npm: ${chalk.green(
+          'my-react-scripts'
+        )}`
+      );
+      console.log(
+        `      - a local path relative to the current working directory: ${chalk.green(
+          'file:../my-react-scripts'
+        )}`
+      );
+      console.log(
+        `      - a .tgz archive: ${chalk.green(
+          'https://mysite.com/my-react-scripts-0.8.2.tgz'
+        )}`
+      );
+      console.log(
+        `      - a .tar.gz archive: ${chalk.green(
+          'https://mysite.com/my-react-scripts-0.8.2.tar.gz'
+        )}`
+      );
+      console.log(
+        `    It is not needed unless you specifically want to use a fork.`
+      );
+      console.log();
+      console.log(`    A custom ${chalk.cyan('--template')} can be one of:`);
+      console.log(
+        `      - a custom template published on npm: ${chalk.green(
+          'cra-template-typescript'
+        )}`
+      );
+      console.log(
+        `      - a local path relative to the current working directory: ${chalk.green(
+          'file:../my-custom-template'
+        )}`
+      );
+      console.log(
+        `      - a .tgz archive: ${chalk.green(
+          'https://mysite.com/my-custom-template-0.8.2.tgz'
+        )}`
+      );
+      console.log(
+        `      - a .tar.gz archive: ${chalk.green(
+          'https://mysite.com/my-custom-template-0.8.2.tar.gz'
+        )}`
+      );
+      console.log();
+      console.log(
+        `    If you have any problems, do not hesitate to file an issue:`
+      );
+      console.log(
+        `      ${chalk.cyan(
+          'https://github.com/facebook/create-react-app/issues/new'
+        )}`
+      );
+      console.log();
+    })
+    .parse(process.argv);
+
+  if (program.info) {
+    console.log(chalk.bold('\nEnvironment Info:'));
+    console.log(
+      `\n  current version of ${packageJson.name}: ${packageJson.version}`
+    );
+    console.log(`  running from ${__dirname}`);
+    return envinfo
+      .run(
+        {
+          System: ['OS', 'CPU'],
+          Binaries: ['Node', 'npm', 'Yarn'],
+          Browsers: [
+            'Chrome',
+            'Edge',
+            'Internet Explorer',
+            'Firefox',
+            'Safari',
+          ],
+          npmPackages: ['react', 'react-dom', 'react-scripts'],
+          npmGlobalPackages: ['create-react-app'],
+        },
+        {
+          duplicates: true,
+          showNotFound: true,
+        }
+      )
+      .then(console.log);
+  }
+
+  if (typeof projectName === 'undefined') {
+    console.error('Please specify the project directory:');
+    console.log(
+      `  ${chalk.cyan(program.name())} ${chalk.green('<project-directory>')}`
+    );
+    console.log();
+    console.log('For example:');
+    console.log(
+      `  ${chalk.cyan(program.name())} ${chalk.green('my-react-app')}`
+    );
+    console.log();
+    console.log(
+      `Run ${chalk.cyan(`${program.name()} --help`)} to see all options.`
+    );
+    process.exit(1);
+  }
+
+  // We first check the registry directly via the API, and if that fails, we try
+  // the slower `npm view [package] version` command.
+  //
+  // This is important for users in environments where direct access to npm is
+  // blocked by a firewall, and packages are provided exclusively via a private
+  // registry.
+  checkForLatestVersion()
+    .catch(() => {
+      try {
+        return execSync('npm view create-react-app version').toString().trim();
+      } catch (e) {
+        return null;
+      }
+    })
+    .then(latest => {
+      if (latest && semver.lt(packageJson.version, latest)) {
+        console.log();
+        console.error(
+          chalk.yellow(
+            `You are running \`create-react-app\` ${packageJson.version}, which is behind the latest release (${latest}).\n\n` +
+              'We recommend always using the latest version of create-react-app if possible.'
+          )
+        );
+        console.log();
+        console.log(
+          'The latest instructions for creating a new app can be found here:\n' +
+            'https://create-react-app.dev/docs/getting-started/'
+        );
+        console.log();
+      } else {
+        const useYarn = isUsingYarn();
+        createApp(
+          projectName,
+          program.verbose,
+          program.scriptsVersion,
+          program.template,
+          useYarn,
+          program.usePnp
+        );
+      }
+    });
 }
 
-if (typeof projectName === 'undefined') {
-  console.error('Please specify the project directory:');
-  console.log(
-    `  ${chalk.cyan(program.name())} ${chalk.green('<project-directory>')}`
+function createApp(name, verbose, version, template, useYarn, usePnp) {
+  const unsupportedNodeVersion = !semver.satisfies(
+    // Coerce strings with metadata (i.e. `15.0.0-nightly`).
+    semver.coerce(process.version),
+    '>=14'
   );
-  console.log();
-  console.log('For example:');
-  console.log(`  ${chalk.cyan(program.name())} ${chalk.green('my-react-app')}`);
-  console.log();
-  console.log(
-    `Run ${chalk.cyan(`${program.name()} --help`)} to see all options.`
-  );
-  process.exit(1);
-}
 
-createApp(
-  projectName,
-  program.verbose,
-  program.scriptsVersion,
-  program.template,
-  program.useNpm,
-  program.usePnp
-);
-
-function createApp(name, verbose, version, template, useNpm, usePnp) {
-  const unsupportedNodeVersion = !semver.satisfies(process.version, '>=10');
   if (unsupportedNodeVersion) {
     console.log(
       chalk.yellow(
         `You are using Node ${process.version} so the project will be bootstrapped with an old unsupported version of tools.\n\n` +
-          `Please update to Node 10 or higher for a better, fully supported experience.\n`
+          `Please update to Node 14 or higher for a better, fully supported experience.\n`
       )
     );
     // Fall back to latest supported react-scripts on Node 4
@@ -224,7 +273,6 @@ function createApp(name, verbose, version, template, useNpm, usePnp) {
     JSON.stringify(packageJson, null, 2) + os.EOL
   );
 
-  const useYarn = useNpm ? false : shouldUseYarn();
   const originalDirectory = process.cwd();
   process.chdir(root);
   if (!useYarn && !checkThatNpmCanReadCwd()) {
@@ -270,23 +318,6 @@ function createApp(name, verbose, version, template, useNpm, usePnp) {
     }
   }
 
-  if (useYarn) {
-    let yarnUsesDefaultRegistry = true;
-    try {
-      yarnUsesDefaultRegistry =
-        execSync('yarnpkg config get registry').toString().trim() ===
-        'https://registry.yarnpkg.com';
-    } catch (e) {
-      // ignore
-    }
-    if (yarnUsesDefaultRegistry) {
-      fs.copySync(
-        require.resolve('./yarn.lock.cached'),
-        path.join(root, 'yarn.lock')
-      );
-    }
-  }
-
   run(
     root,
     appName,
@@ -297,15 +328,6 @@ function createApp(name, verbose, version, template, useNpm, usePnp) {
     useYarn,
     usePnp
   );
-}
-
-function shouldUseYarn() {
-  try {
-    execSync('yarnpkg --version', { stdio: 'ignore' });
-    return true;
-  } catch (e) {
-    return false;
-  }
 }
 
 function install(root, useYarn, usePnp, dependencies, verbose, isOnline) {
@@ -340,6 +362,7 @@ function install(root, useYarn, usePnp, dependencies, verbose, isOnline) {
       command = 'npm';
       args = [
         'install',
+        '--no-audit', // https://github.com/facebook/create-react-app/issues/11174
         '--save',
         '--save-exact',
         '--loglevel',
@@ -465,7 +488,7 @@ function run(
           },
           [root, appName, verbose, originalDirectory, templateName],
           `
-        var init = require('${packageName}/scripts/init.js');
+        const init = require('${packageName}/scripts/init.js');
         init.apply(null, JSON.parse(process.argv[1]));
       `
         );
@@ -474,7 +497,7 @@ function run(
           console.log(
             chalk.yellow(
               `\nNote: the project was bootstrapped with an old unsupported version of tools.\n` +
-                `Please update to Node >=10 and npm >=6 to get supported tools in new projects.\n`
+                `Please update to Node >=14 and npm >=6 to get supported tools in new projects.\n`
             )
           );
         }
@@ -493,11 +516,7 @@ function run(
         console.log();
 
         // On 'exit' we will delete these files from target directory.
-        const knownGeneratedFiles = [
-          'package.json',
-          'yarn.lock',
-          'node_modules',
-        ];
+        const knownGeneratedFiles = ['package.json', 'node_modules'];
         const currentFiles = fs.readdirSync(path.join(root));
         currentFiles.forEach(file => {
           knownGeneratedFiles.forEach(fileToMatch => {
@@ -557,20 +576,18 @@ function getInstallPackage(version, originalDirectory) {
 
   for (const script of scriptsToWarn) {
     if (packageToInstall.startsWith(script.name)) {
-      return inquirer
-        .prompt({
-          type: 'confirm',
-          name: 'useScript',
-          message: script.message,
-          default: false,
-        })
-        .then(answer => {
-          if (!answer.useScript) {
-            process.exit(0);
-          }
+      return prompts({
+        type: 'confirm',
+        name: 'useScript',
+        message: script.message,
+        initial: false,
+      }).then(answer => {
+        if (!answer.useScript) {
+          process.exit(0);
+        }
 
-          return packageToInstall;
-        });
+        return packageToInstall;
+      });
     }
   }
 
@@ -593,10 +610,11 @@ function getTemplateInstallPackage(template, originalDirectory) {
       templateToInstall = template;
     } else {
       // Add prefix 'cra-template-' to non-prefixed templates, leaving any
-      // @scope/ intact.
-      const packageMatch = template.match(/^(@[^/]+\/)?(.+)$/);
+      // @scope/ and @version intact.
+      const packageMatch = template.match(/^(@[^/]+\/)?([^@]+)?(@.+)?$/);
       const scope = packageMatch[1] || '';
-      const templateName = packageMatch[2];
+      const templateName = packageMatch[2] || '';
+      const version = packageMatch[3] || '';
 
       if (
         templateName === templateToInstall ||
@@ -607,15 +625,15 @@ function getTemplateInstallPackage(template, originalDirectory) {
         // - @SCOPE/cra-template
         // - cra-template-NAME
         // - @SCOPE/cra-template-NAME
-        templateToInstall = `${scope}${templateName}`;
-      } else if (templateName.startsWith('@')) {
+        templateToInstall = `${scope}${templateName}${version}`;
+      } else if (version && !scope && !templateName) {
         // Covers using @SCOPE only
-        templateToInstall = `${templateName}/${templateToInstall}`;
+        templateToInstall = `${version}/${templateToInstall}`;
       } else {
         // Covers templates without the `cra-template` prefix:
         // - NAME
         // - @SCOPE/NAME
-        templateToInstall = `${scope}${templateToInstall}-${templateName}`;
+        templateToInstall = `${scope}${templateToInstall}-${templateName}${version}`;
       }
     }
   }
@@ -1075,3 +1093,31 @@ function executeNodeScript({ cwd, args }, data, source) {
     });
   });
 }
+
+function checkForLatestVersion() {
+  return new Promise((resolve, reject) => {
+    https
+      .get(
+        'https://registry.npmjs.org/-/package/create-react-app/dist-tags',
+        res => {
+          if (res.statusCode === 200) {
+            let body = '';
+            res.on('data', data => (body += data));
+            res.on('end', () => {
+              resolve(JSON.parse(body).latest);
+            });
+          } else {
+            reject();
+          }
+        }
+      )
+      .on('error', () => {
+        reject();
+      });
+  });
+}
+
+module.exports = {
+  init,
+  getTemplateInstallPackage,
+};
