@@ -36,19 +36,21 @@
 
 'use strict';
 
-var currentNodeVersion = process.versions.node;
-var semver = currentNodeVersion.split('.');
-var major = semver[0];
+const currentNodeVersion = process.versions.node;
+const semver = currentNodeVersion.split('.');
+const major = semver[0];
 
-if (major < 10) {
+if (major < 14) {
   console.error(
     'You are running Node ' +
       currentNodeVersion +
       '.\n' +
-      'Create React App requires Node 10 or higher. \n' +
+      'Create React App requires Node 14 or higher. \n' +
       'Please update your version of Node.'
   );
   process.exit(1);
 }
 
-require('./createReactApp');
+const { init } = require('./createReactApp');
+
+init();
