@@ -7,15 +7,26 @@
 
 'use strict';
 
+const path = require('path');
 const chalk = require('chalk');
 const stripAnsi = require('strip-ansi');
 const table = require('text-table');
+
+const cwd = process.cwd();
+
+const emitErrorsAsWarnings =
+  process.env.NODE_ENV === 'development' &&
+  process.env.ESLINT_NO_DEV_ERRORS === 'true';
 
 function isError(message) {
   if (message.fatal || message.severity === 2) {
     return true;
   }
   return false;
+}
+
+function getRelativePath(filePath) {
+  return path.relative(cwd, filePath);
 }
 
 function formatter(results) {
@@ -31,7 +42,7 @@ function formatter(results) {
 
     messages = messages.map(message => {
       let messageType;
-      if (isError(message)) {
+      if (isError(message) && !emitErrorsAsWarnings) {
         messageType = 'error';
         hasErrors = true;
         if (message.ruleId) {
@@ -73,6 +84,10 @@ function formatter(results) {
       },
     });
 
+    // print the filename and relative path
+    output += `${getRelativePath(result.filePath)}\n`;
+
+    // print the errors
     output += `${outputTable}\n\n`;
   });
 
