@@ -45,7 +45,7 @@ function getBrowserEnv() {
 
 function executeNodeScript(scriptPath, url) {
   const extraArgs = process.argv.slice(2);
-  const child = spawn('node', [scriptPath, ...extraArgs, url], {
+  const child = spawn(process.execPath, [scriptPath, ...extraArgs, url], {
     stdio: 'inherit',
   });
   child.on('close', code => {
@@ -77,6 +77,8 @@ function startBrowserProcess(browser, url, args) {
     // Will use the first open browser found from list
     const supportedChromiumBrowsers = [
       'Google Chrome Canary',
+      'Google Chrome Dev',
+      'Google Chrome Beta',
       'Google Chrome',
       'Microsoft Edge',
       'Brave Browser',
