@@ -51,7 +51,7 @@ export function reportBuildError(error: string) {
 
 export function reportRuntimeError(
   error: Error,
-  options?: RuntimeReportingOption = {}
+  options: RuntimeReportingOptions = {}
 ) {
   currentRuntimeErrorOptions = options;
   crashWithFrames(handleRuntimeError(options))(error);
@@ -80,29 +80,29 @@ export function startReportingRuntimeErrors(options: RuntimeReportingOptions) {
   );
 }
 
-const handleRuntimeError = (options: RuntimeReportingOptions) => (
-  errorRecord: ErrorRecord
-) => {
-  try {
-    if (typeof options.onError === 'function') {
-      options.onError.call(null);
+const handleRuntimeError =
+  (options: RuntimeReportingOptions) => (errorRecord: ErrorRecord) => {
+    try {
+      if (typeof options.onError === 'function') {
+        options.onError.call(null);
+      }
+    } finally {
+      if (
+        currentRuntimeErrorRecords.some(
+          ({ error }) => error === errorRecord.error
+        )
+      ) {
+        // Deduplicate identical errors.
+        // This fixes https://github.com/facebook/create-react-app/issues/3011.
+        // eslint-disable-next-line no-unsafe-finally
+        return;
+      }
+      currentRuntimeErrorRecords = currentRuntimeErrorRecords.concat([
+        errorRecord,
+      ]);
+      update();
     }
-  } finally {
-    if (
-      currentRuntimeErrorRecords.some(
-        ({ error }) => error === errorRecord.error
-      )
-    ) {
-      // Deduplicate identical errors.
-      // This fixes https://github.com/facebook/create-react-app/issues/3011.
-      return;
-    }
-    currentRuntimeErrorRecords = currentRuntimeErrorRecords.concat([
-      errorRecord,
-    ]);
-    update();
-  }
-};
+  };
 
 export function dismissRuntimeErrors() {
   currentRuntimeErrorRecords = [];
@@ -142,9 +142,8 @@ function update() {
     const iframeDocument = loadingIframe.contentDocument;
     if (iframeDocument != null && iframeDocument.body != null) {
       iframe = loadingIframe;
-      const script = loadingIframe.contentWindow.document.createElement(
-        'script'
-      );
+      const script =
+        loadingIframe.contentWindow.document.createElement('script');
       script.type = 'text/javascript';
       script.innerHTML = iframeScript;
       iframeDocument.body.appendChild(script);
@@ -179,11 +178,12 @@ function updateIframeContent() {
 
 window.__REACT_ERROR_OVERLAY_GLOBAL_HOOK__ =
   window.__REACT_ERROR_OVERLAY_GLOBAL_HOOK__ || {};
-window.__REACT_ERROR_OVERLAY_GLOBAL_HOOK__.iframeReady = function iframeReady() {
-  isIframeReady = true;
-  isLoadingIframe = false;
-  updateIframeContent();
-};
+window.__REACT_ERROR_OVERLAY_GLOBAL_HOOK__.iframeReady =
+  function iframeReady() {
+    isIframeReady = true;
+    isLoadingIframe = false;
+    updateIframeContent();
+  };
 
 if (process.env.NODE_ENV === 'production') {
   console.warn(
